+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## Major changes
 
+- The `count` outcome type has been deprecated. `count` is a subset of `continuous` outcomes. Its previous implementation did not provide any benefits over `continuous`.
+
 - Gradient boosting using the `mboost` package was deprecated. Use `xgboost` instead.
 
 - The `qvalue` package for computing q-values was deprecated.